[common/meta/types] refactor: move TableInfo into common/meta/types
- fix: #2138
