+++ common/meta/flight/src/meta_api.rs
@@ -16,6 +16,7 @@
 use common_exception::Result;
 use common_meta_types::MetaId;
 use common_meta_types::MetaVersion;
+use common_meta_types::TableInfo;
 use common_planners::CreateDatabasePlan;
 use common_planners::CreateTablePlan;
 use common_planners::DropDatabasePlan;
@@ -26,7 +27,6 @@ use crate::meta_flight_reply::CreateTableReply;
 use crate::meta_flight_reply::DatabaseInfo;
 use crate::meta_flight_reply::GetDatabasesReply;
 use crate::meta_flight_reply::GetTablesReply;
-use crate::meta_flight_reply::TableInfo;
 
 #[async_trait::async_trait]
 pub trait MetaApi: Send + Sync {
