+++ query/src/catalogs/impls/meta_backends/embedded_meta_backend.rs
@@ -21,9 +21,9 @@ use common_infallible::RwLock;
 use common_meta_flight::meta_flight_reply::CreateDatabaseReply;
 use common_meta_flight::meta_flight_reply::CreateTableReply;
 use common_meta_flight::meta_flight_reply::DatabaseInfo;
-use common_meta_flight::meta_flight_reply::TableInfo;
 use common_meta_types::MetaId;
 use common_meta_types::MetaVersion;
+use common_meta_types::TableInfo;
 use common_planners::CreateDatabasePlan;
 use common_planners::CreateTablePlan;
 use common_planners::DropDatabasePlan;
