+++ query/src/datasources/table/memory/memory_table.rs
@@ -23,7 +23,7 @@ use common_datavalues::DataSchemaRef;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_infallible::RwLock;
-use common_meta_flight::meta_flight_reply::TableInfo;
+use common_meta_types::TableInfo;
 use common_planners::Extras;
 use common_planners::InsertIntoPlan;
 use common_planners::ReadDataSourcePlan;
