+++ query/src/datasources/table/parquet/parquet_table_test.rs
@@ -19,7 +19,7 @@ use std::sync::Arc;
 use common_base::tokio;
 use common_datavalues::prelude::*;
 use common_exception::Result;
-use common_meta_flight::meta_flight_reply::TableInfo;
+use common_meta_types::TableInfo;
 use common_planners::*;
 use futures::TryStreamExt;
