+++ query/src/datasources/table_engine.rs
@@ -13,7 +13,7 @@
 //  limitations under the License.
 //
 
-use common_meta_flight::meta_flight_reply::TableInfo;
+use common_meta_types::TableInfo;
 
 use crate::catalogs::Table;
 use crate::common::StoreApiProvider;
